Changes in multiple areas. Added a playstate class which doesn't do anything right now...
We gotta make a change...
It's time for us as a people to start makin' some changes.
Let's change the way we eat, let's change the way we live
and let's change the way we treat each other.
You see the old way wasn't working so it's on us to do
what we gotta do, to survive.